+++ langchain/agents/agent_toolkits/spark_sql/toolkit.py
@@ -5,7 +5,6 @@ from pydantic import Field
 
 from langchain.agents.agent_toolkits.base import BaseToolkit
 from langchain.base_language import BaseLanguageModel
-from langchain.spark_sql import SparkSQL
 from langchain.tools import BaseTool
 from langchain.tools.spark_sql.tool import (
     InfoSparkSQLTool,
@@ -13,6 +12,7 @@ from langchain.tools.spark_sql.tool import (
     QueryCheckerTool,
     QuerySparkSQLTool,
 )
+from langchain.utilities.spark_sql import SparkSQL
 
 
 class SparkSQLToolkit(BaseToolkit):
