+++ langchain/tools/spark_sql/tool.py
@@ -11,7 +11,7 @@ from langchain.callbacks.manager import (
 )
 from langchain.chains.llm import LLMChain
 from langchain.prompts import PromptTemplate
-from langchain.spark_sql import SparkSQL
+from langchain.utilities.spark_sql import SparkSQL
 from langchain.tools.base import BaseTool
 from langchain.tools.spark_sql.prompt import QUERY_CHECKER
